frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.39.0"
requests = "^2.32.3"
python-dotenv = "^0.21.0" 
boto3 = "^1.35.49"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
